Add check-msvc warning sweep and Windows headless smoke coverage
Local MSVC-mirror gate (make check-msvc, part of verify-all) builds the tree with Clang -Wfloat-conversion/-Wshorten-64-to-32/-Wshadow-field so the Windows-only C4244/C4267/C4458 classes fail before CI, fixing six engine int-to-float casts, two member shadows, and the title-to-windowTitle ctor-param idiom. The Windows CI job now smoke-runs every example and game headless (SDL dummy drivers); smoke.cmake drops GNU timeout for CMake native execute_process TIMEOUT so the shared acceptance rule is cross-platform, and smoke_* ctest entries register on MSVC with a ctest -N guard.

+++ Engine/Core/FPSCounter.h
@@ -15,7 +15,7 @@ public:
         Uint32 current_time = SDL_GetTicks();
 
         if (current_time - start_time >= 500) {  // Update every 500 ms
-            fps = (frame_count * 1000.0f) / (current_time - start_time);
+            fps = (static_cast<float>(frame_count) * 1000.0f) / static_cast<float>(current_time - start_time);
             frame_count = 0;
             start_time = current_time;
         }
